MCP_EdgeMPSoc/requirements.txt
# WattMCP Dependencies

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# MQTT
paho-mqtt==1.6.1

# Redis for caching
redis==5.0.1

# HTTP client for API calls
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Optional dependencies for enhanced functionality
# - TLS/SSL support
# - Database support
# - Monitoring
prometheus-client==0.19.0